requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-auth==2.35.0
google-api-python-client==2.149.0
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.47.0
openai==1.50.2